# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Narrative Simulation Engine API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type WorldStateActiveMode = typeof WorldStateActiveMode[keyof typeof WorldStateActiveMode];


export const WorldStateActiveMode = {
  novel: 'novel',
  rpg: 'rpg',
  advanced: 'advanced',
  auto: 'auto',
} as const;

export interface WorldState {
  id: number;
  /** Current in-game day number */
  worldDay: number;
  /** Current in-game time (HH:MM) */
  worldTime: string;
  /** Current weather condition */
  weather: string;
  currentLocationId: number;
  currentLocationName: string;
  activeMode: WorldStateActiveMode;
  isRunning: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface WorldTickInput {
  /** How many in-game minutes to advance */
  minutesToAdvance?: number;
}

export type WorldEventType = typeof WorldEventType[keyof typeof WorldEventType];


export const WorldEventType = {
  scheduled: 'scheduled',
  random: 'random',
  plot: 'plot',
  social: 'social',
  consequence: 'consequence',
  npc_action: 'npc_action',
  player_action: 'player_action',
} as const;

export interface WorldEvent {
  id: number;
  type: WorldEventType;
  title: string;
  description: string;
  isActive: boolean;
  worldDay: number;
  worldTime: string;
  locationId: number;
  locationName: string;
  involvedNpcIds?: number[];
  createdAt: string;
}

export interface NpcAction {
  npcId: number;
  npcName: string;
  action: string;
  locationId: number;
  locationName: string;
  /** @nullable */
  narrative?: string | null;
}

export interface WorldTickResult {
  worldState: WorldState;
  triggeredEvents: WorldEvent[];
  npcActions: NpcAction[];
  /**
     * Short narrative description of what happened during the tick
     * @nullable
     */
  narrativeSummary?: string | null;
}

export interface NpcSummary {
  id: number;
  name: string;
  locationId: number;
  locationName: string;
  currentActivity: string;
  emotionalState: string;
  /** @nullable */
  relationshipWithPlayer?: string | null;
}

export type MemoryEntryOwnerType = typeof MemoryEntryOwnerType[keyof typeof MemoryEntryOwnerType];


export const MemoryEntryOwnerType = {
  player: 'player',
  npc: 'npc',
  world: 'world',
} as const;

export type MemoryEntryImportance = typeof MemoryEntryImportance[keyof typeof MemoryEntryImportance];


export const MemoryEntryImportance = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface MemoryEntry {
  id: number;
  /** Who owns this memory (NPC id or 0 for world/player) */
  ownerId: number;
  ownerType: MemoryEntryOwnerType;
  /**
     * NPC this memory is about (if any)
     * @nullable
     */
  subjectId?: number | null;
  /** @nullable */
  subjectName?: string | null;
  /** What is remembered */
  content: string;
  importance: MemoryEntryImportance;
  worldDay: number;
  worldTime: string;
  /** Long-term memory (persists) vs short-term (may fade) */
  isLongTerm: boolean;
  createdAt: string;
}

export interface WorldSummary {
  worldState: WorldState;
  presentNpcs: NpcSummary[];
  activeEvents: WorldEvent[];
  recentMemories: MemoryEntry[];
  relationshipCount: number;
}

export interface Location {
  id: number;
  name: string;
  description: string;
  /** indoor, outdoor, settlement, wilderness, etc. */
  type: string;
  isAccessible: boolean;
  /** @nullable */
  npcCount?: number | null;
}

export interface LocationDetail {
  id: number;
  name: string;
  description: string;
  type: string;
  isAccessible: boolean;
  presentNpcs: NpcSummary[];
  recentEvents: WorldEvent[];
}

export interface NpcDetail {
  id: number;
  name: string;
  /** @nullable */
  age: number | null;
  /** Short description of NPC personality */
  personality: string;
  /** @nullable */
  background?: string | null;
  objectives: string[];
  /** What the NPC is currently doing */
  currentRoutine: string;
  locationId: number;
  locationName: string;
  emotionalState: string;
  /** Secrets this NPC holds (visible only in advanced mode) */
  knownSecrets: string[];
  createdAt: string;
}

export interface EventPage {
  events: WorldEvent[];
  total: number;
  offset: number;
  limit: number;
}

export interface PlayerActionInput {
  /**
     * The player's action or dialogue input
     * @minLength 1
     */
  text: string;
  /**
     * NPC this action is directed at (if any)
     * @nullable
     */
  targetNpcId?: number | null;
}

export interface NarrativeResponse {
  /** The narrative text response */
  narrative: string;
  worldStateChanged: boolean;
  newEvents: WorldEvent[];
  npcReactions?: NpcAction[];
  newMemories?: MemoryEntry[];
  messageId: number;
  worldDay: number;
  worldTime: string;
}

export type NarrativeMessageRole = typeof NarrativeMessageRole[keyof typeof NarrativeMessageRole];


export const NarrativeMessageRole = {
  player: 'player',
  narrator: 'narrator',
} as const;

export interface NarrativeMessage {
  id: number;
  role: NarrativeMessageRole;
  text: string;
  worldDay: number;
  worldTime: string;
  createdAt: string;
}

export interface Relationship {
  id: number;
  npcId: number;
  npcName: string;
  /**
     * @minimum -100
     * @maximum 100
     */
  trust: number;
  /**
     * @minimum -100
     * @maximum 100
     */
  respect: number;
  /**
     * @minimum -100
     * @maximum 100
     */
  suspicion: number;
  /**
     * @minimum -100
     * @maximum 100
     */
  friendship: number;
  /**
     * @minimum -100
     * @maximum 100
     */
  rivalry: number;
  /** Narrative label (stranger, acquaintance, ally, enemy, etc.) */
  status: string;
  /** @nullable */
  lastInteractionDay: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface SaveSlot {
  id: number;
  name: string;
  worldDay: number;
  worldTime: string;
  locationName: string;
  createdAt: string;
}

export interface SaveInput {
  /** @minLength 1 */
  name: string;
}

export type GameSettingsActiveMode = typeof GameSettingsActiveMode[keyof typeof GameSettingsActiveMode];


export const GameSettingsActiveMode = {
  novel: 'novel',
  rpg: 'rpg',
  advanced: 'advanced',
  auto: 'auto',
} as const;

export type GameSettingsNarrativeSpeed = typeof GameSettingsNarrativeSpeed[keyof typeof GameSettingsNarrativeSpeed];


export const GameSettingsNarrativeSpeed = {
  slow: 'slow',
  normal: 'normal',
  fast: 'fast',
} as const;

export interface GameSettings {
  id: number;
  activeMode: GameSettingsActiveMode;
  narrativeSpeed: GameSettingsNarrativeSpeed;
  autoTickEnabled: boolean;
  autoTickIntervalMinutes: number;
  language: string;
  updatedAt: string;
}

export type GameSettingsInputActiveMode = typeof GameSettingsInputActiveMode[keyof typeof GameSettingsInputActiveMode];


export const GameSettingsInputActiveMode = {
  novel: 'novel',
  rpg: 'rpg',
  advanced: 'advanced',
  auto: 'auto',
} as const;

export type GameSettingsInputNarrativeSpeed = typeof GameSettingsInputNarrativeSpeed[keyof typeof GameSettingsInputNarrativeSpeed];


export const GameSettingsInputNarrativeSpeed = {
  slow: 'slow',
  normal: 'normal',
  fast: 'fast',
} as const;

export interface GameSettingsInput {
  activeMode?: GameSettingsInputActiveMode;
  narrativeSpeed?: GameSettingsInputNarrativeSpeed;
  autoTickEnabled?: boolean;
  autoTickIntervalMinutes?: number;
  language?: string;
}

export type ListEventsParams = {
limit?: number;
offset?: number;
type?: string;
};

export type ListMemoriesParams = {
limit?: number;
/**
 * Filter by subject NPC id
 * @nullable
 */
subjectId?: number | null;
};

export type GetNarrativeHistoryParams = {
limit?: number;
};