# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Narrative Simulation Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /world/tick:
    post:
      operationId: tickWorld
      tags: [world]
      summary: Advance world time
      description: Moves time forward, triggers NPC routines and autonomous events
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorldTickInput"
      responses:
        "200":
          description: Updated world state after tick
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorldTickResult"
...
  /world/locations/{locationId}:
    get:
      operationId: getLocation
      tags: [world]
      summary: Get a location with present NPCs
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Location details with present NPCs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationDetail"
        "404":
          description: Location not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── NPC ENGINE ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ─── WORLD ────────────────────────────────────────────────────────────────
    WorldState:
      type: object
      required: [id, worldDay, worldTime, weather, currentLocationId, currentLocationName, activeMode, isRunning, createdAt, updatedAt]
      properties:
        id:
          type: number
        worldDay:
          type: number
          description: Current in-game day number
        worldTime:
          type: string
          description: Current in-game time (HH:MM)
        weather:
          type: string
          description: Current weather condition
        currentLocationId:
          type: number
        currentLocationName:
          type: string
        activeMode:
          type: string
          enum: [novel, rpg, advanced, auto]
        isRunning:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WorldTickInput:
      type: object
      properties:
        minutesToAdvance:
          type: number
          default: 60
          description: How many in-game minutes to advance
    WorldTickResult:
      type: object
      required: [worldState, triggeredEvents, npcActions]
      properties:
        worldState:
          $ref: "#/components/schemas/WorldState"
        triggeredEvents:
          type: array
          items:
            $ref: "#/components/schemas/WorldEvent"
        npcActions:
          type: array
          items:
            $ref: "#/components/schemas/NpcAction"
        narrativeSummary:
          type: ["string", "null"]
          description: Short narrative description of what happened during the tick
...
    LocationDetail:
      type: object
      required: [id, name, description, type, isAccessible, presentNpcs, recentEvents]
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: string
        type:
          type: string
        isAccessible:
          type: boolean
        presentNpcs:
          type: array
          items:
            $ref: "#/components/schemas/NpcSummary"
        recentEvents:
          type: array
          items:
            $ref: "#/components/schemas/WorldEvent"

    # ─── NPC ──────────────────────────────────────────────────────────────────
    NpcSummary:
      type: object
      required: [id, name, locationId, locationName, currentActivity, emotionalState]
      properties:
        id:
          type: number
        name:
          type: string
        locationId:
          type: number
        locationName:
          type: string
        currentActivity:
          type: string
        emotionalState:
          type: string
        relationshipWithPlayer:
          type: ["string", "null"]
...
    NpcAction:
      type: object
      required: [npcId, npcName, action, locationId, locationName]
      properties:
        npcId:
          type: number
        npcName:
          type: string
        action:
          type: string
        locationId:
          type: number
        locationName:
          type: string
        narrative:
          type: ["string", "null"]

    # ─── EVENTS ───────────────────────────────────────────────────────────────
    WorldEvent:
      type: object
      required: [id, type, title, description, isActive, worldDay, worldTime, locationId, locationName, createdAt]
      properties:
        id:
          type: number
        type:
          type: string
          enum: [scheduled, random, plot, social, consequence, npc_action, player_action]
        title:
          type: string
        description:
          type: string
        isActive:
          type: boolean
        worldDay:
          type: number
        worldTime:
          type: string
        locationId:
          type: number
        locationName:
          type: string
        involvedNpcIds:
          type: array
          items:
            type: number
        createdAt:
          type: string
          format: date-time
...