# lib/api-zod/src/generated/types/npcAction.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Narrative Simulation Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface NpcAction {
  npcId: number;
  npcName: string;
  action: string;
  locationId: number;
  locationName: string;
  /** @nullable */
  narrative?: string | null;
}